requirements.txt
opencv_python==4.11.0.86
pycryptodome==3.23.0
pyotp==2.9.0
PySide6==6.9.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-dotenv==1.1.1
pyzbar==0.1.9
